+++ lib/src/televerse/models/models.dart
@@ -10,7 +10,7 @@ import 'package:televerse/televerse.dart';
 part 'chat_id.dart';
 part 'input_file.dart';
 part 'webhook_exception.dart';
-part 'http_exception.dart';
+part 'telegram_exception.dart';
 part 'long_polling_exception.dart';
 part 'televerse_exception.dart';
 part 'mention.dart';