+++ lib/src/utils/http.dart
@@ -7,17 +7,12 @@ import 'package:televerse/televerse.dart';
 class HttpClient {
   /// Send GET request to the given [uri] and return the response body.
   static Future<dynamic> getURI(Uri uri) async {
-    try {
-      final response = await get(uri);
-      final body = json.decode(response.body);
-      if (body["ok"] == true) {
-        return body["result"];
-      } else {
-        final ex = HttpException(response.statusCode, body["description"]);
-        _handleException(ex.code, ex.message);
-      }
-    } catch (err) {
-      return Future.error(err);
+    final response = await get(uri);
+    final body = json.decode(response.body);
+    if (body["ok"] == true) {
+      return body["result"];
+    } else {
+      throw TelegramException.fromJson(body);
     }
   }
 
@@ -33,16 +28,13 @@ class HttpClient {
       }
       return MapEntry(k, "$v");
     });
-    try {
-      final response = await post(uri, body: bodyContent);
-      final resBody = json.decode(response.body);
-      if (resBody["ok"] == true) {
-        return resBody["result"];
-      } else {
-        throw HttpException(response.statusCode, resBody["description"]);
-      }
-    } catch (err) {
-      return Future.error(err);
+
+    final response = await post(uri, body: bodyContent);
+    final resBody = json.decode(response.body);
+    if (resBody["ok"] == true) {
+      return resBody["result"];
+    } else {
+      throw TelegramException.fromJson(resBody);
     }
   }
 
@@ -52,33 +44,18 @@ class HttpClient {
     List<MultipartFile> files,
     Map<String, dynamic> body,
   ) async {
-    try {
-      body.removeWhere((key, value) => value == null || value == "null");
-      final request = MultipartRequest("POST", uri)
-        ..headers.addAll({"Content-Type": "multipart/form-data"})
-        ..fields.addAll(body.map((k, v) => MapEntry(k, "$v")))
-        ..files.addAll(files);
-      final response = await request.send();
-      final resBody = await response.stream.bytesToString();
-      final res = json.decode(resBody);
-      if (res["ok"] == true) {
-        return res["result"];
-      } else {
-        throw HttpException(response.statusCode, res["description"]);
-      }
-    } catch (err) {
-      return Future.error(err);
-    }
-  }
-
-  /// Handle the given [code] and [message].
-  static void _handleException(int code, String message) {
-    if (code == 401) {
-      throw UnauthorizedException(message);
-    } else if (code == 400) {
-      throw BadRequestException(message);
+    body.removeWhere((key, value) => value == null || value == "null");
+    final request = MultipartRequest("POST", uri)
+      ..headers.addAll({"Content-Type": "multipart/form-data"})
+      ..fields.addAll(body.map((k, v) => MapEntry(k, "$v")))
+      ..files.addAll(files);
+    final response = await request.send();
+    final resBody = await response.stream.bytesToString();
+    final res = json.decode(resBody);
+    if (res["ok"] == true) {
+      return res["result"];
     } else {
-      throw HttpException(code, message);
+      throw TelegramException.fromJson(res);
     }
   }
 }